refactor: remove unsafe cast and unify bindings naming
- Fix useViewerController fetcher type to eliminate unsafe `as never` cast
- Rename BookshelfCommand.ts, SeriesCommand.ts to plural form (*Commands.ts) for consistency with existing conventions
- Update all import references and mocks across the codebase
