+++ src/features/Bookshelf/components/Dialog/AddBooksToBookshelvesDialog.tsx
@@ -15,7 +15,7 @@ import {
 import { error as logError } from "@tauri-apps/plugin-log";
 import { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { addBookToBookshelf } from "../../../../bindings/BookshelfCommand";
+import { addBookToBookshelf } from "../../../../bindings/BookshelfCommands";
 import type { Bookshelf } from "../../../../domain/bookshelf/schema";
 import { BookShelfIcons } from "../BookshelfIcons";
