+++ src/features/Bookshelf/components/Dialog/SetSeriesDialog.tsx
@@ -18,7 +18,7 @@ import { error as logError } from "@tauri-apps/plugin-log";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { updateBookSeries } from "../../../../bindings/BookCommands";
-import { createSeries } from "../../../../bindings/SeriesCommand";
+import { createSeries } from "../../../../bindings/SeriesCommands";
 import type { Series } from "../../../../domain/series/schema";
 
 /** Props for the SetSeriesDialog component */
